(split)bug: 22362
Handle MailItems representing items larger than 2GB.  Increase the MAIL_ITEM.SIZE and REVISION.SIZE database columns to BIGINT.  Change MailItem.mData.size and MailItem.getSize() to return a long.  Tweak a few things around the app to accommodate this, with the only sketchy bit being that the methods returning item content in a byte[] now must throw MailServiceException.MESSAGE_TOO_BIG on >2GB items because Java doesn't support byte arrays larger than that.

Add Folder.getItemCount() and try to call it instead of calling Folder.getSize() when we want to know the number of non-subfolder items in a folder.

DB version number is now 50.  Please reset-the-world or run the migrate20071206-WidenSizeColumns.pl migration script to get your install's database schema in sync.

http://bugzilla.zimbra.com/show_bug.cgi?id=22362

Copied from Perforce
 Change: 76304

+++ src/java/com/zimbra/cs/dav/resource/Notebook.java
@@ -51,7 +51,7 @@ public class Notebook extends MailItemResource {
 		setProperty(DavElements.P_DISPLAYNAME, doc.getName());
 		// content length is just an estimate.  the actual content will be larger
 		// after chrome composition.
-		setProperty(DavElements.P_GETCONTENTLENGTH, Integer.toString(doc.getSize()));
+		setProperty(DavElements.P_GETCONTENTLENGTH, Long.toString(doc.getSize()));
 		setProperty(DavElements.P_GETCONTENTTYPE, doc.getContentType());
 	}
